Modify : 헤더 컴포넌트로 state, 함수 이전.
